+++ tests/testing-utils/drawing_utils.js
